requirements.txt
# Core Django and Web Framework
Django>=4.2.0,<5.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0

# Database Support
mysqlclient>=2.1.0

# ChatGLM Integration (OpenAI-compatible API)
openai>=1.0.0

# Hugging Face and Machine Learning
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
huggingface-hub>=0.16.0
tokenizers>=0.13.0
safetensors>=0.3.0

# Vector Search and Similarity
faiss-cpu>=1.7.0
# Note: Use faiss-gpu instead of faiss-cpu if you have CUDA support

# Scientific Computing
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# LangChain for Document Processing
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# HTTP Requests and Network
requests>=2.31.0
urllib3>=1.26.0

# Text Processing (Chinese Support)
jieba>=0.42.1

# Image Processing
Pillow>=10.0.0

# Utilities
python-multipart>=0.0.6
python-dateutil>=2.8.0
pytz>=2023.3
six>=1.16.0

# File Processing
chardet>=5.0.0
charset-normalizer>=3.0.0

# JSON and Data Serialization
ujson>=5.7.0

# Development and Debugging (Optional)
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0

# Production Deployment (Optional)
gunicorn>=21.0.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.0

# Caching (Optional)
redis>=4.5.0
django-redis>=5.2.0

# Monitoring and Logging (Optional)
sentry-sdk>=1.25.0

# Security
cryptography>=41.0.0

# API Documentation (Optional)
djangorestframework>=3.14.0
drf-yasg>=1.21.0

# Testing (Optional)
pytest>=7.3.0
pytest-django>=4.5.0
factory-boy>=3.2.0

# Code Quality (Optional)
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0